huuge commit... added leader 'election', sate replicaiition, and also a client failover
the nodoes elect a leader based on the highest port number. And the leader sends heartbeats to followers every 2 secs. The followers then detect leader failure after 5 secs timeout (of not getting heartbeat). The followers are forawrding bids to the leader through ForwardBid rpc. The leader is replicating state to followers after each and every bid. And as mentioned, the client supports multiple servers with now automatic failover